refactor: add lint and fix most lint issues
