+++ package.json
@@ -7,7 +7,8 @@
     "packages/schemas"
   ],
   "scripts": {
-    "test": "tap --debug"
+    "test": "tap --debug",
+    "lint": "eslint ."
   },
   "dependencies": {
     "node-fetch": "^3.2.7"
@@ -15,6 +16,11 @@
   "devDependencies": {
     "@eik/common-helpers": "file:packages/helpers",
     "@types/node": "^18.0.6",
+    "eslint": "7.32.0",
+    "eslint-config-airbnb-base": "14.2.1",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-prettier": "4.2.1",
     "tap": "^15.2.3"
   }
 }
