+++ packages/classes/package.json
@@ -5,6 +5,9 @@
   "main": "index.js",
   "license": "ISC",
   "dependencies": {
-    "mime-types": "^2.1.35"
+    "mime-types": "^2.1.35",
+    "@eik/common-helpers": "1.0.0",
+    "@eik/common-classes": "1.0.0",
+    "tap": "^15.2.3"
   }
 }
