+++ packages/helpers/package.json
@@ -7,13 +7,16 @@
   "devDependencies": {
     "@hapi/hapi": "^20.2.2",
     "@types/node": "^18.0.3",
-    "express": "^4.18.1",
-    "fastify": "^4.2.1",
-    "node-fetch": "^3.2.8",
+    "express": "4.18.1",
+    "fastify": "3.29.0",
     "stoppable": "^1.1.0"
   },
   "dependencies": {
     "glob": "^8.0.3",
-    "is-glob": "^4.0.3"
+    "node-fetch": "^3.2.8",
+    "is-glob": "^4.0.3",
+    "tap": "^15.2.3",
+    "@eik/common-helpers": "1.0.0",
+    "@eik/common-classes": "1.0.0"
   }
 }
