+++ packages/schemas/package.json
@@ -14,6 +14,8 @@
     "ajv": "^8.11.0",
     "ajv-formats": "^2.1.1",
     "semver": "^7.3.7",
-    "validate-npm-package-name": "^4.0.0"
+    "validate-npm-package-name": "^4.0.0",
+    "tap": "^15.2.3",
+    "@eik/common-schemas": "1.0.0"
   }
 }
