+++ packages/validators/package.json
@@ -9,6 +9,7 @@
   "license": "ISC",
   "dependencies": {
     "semver": "^7.3.7",
-    "validate-npm-package-name": "^4.0.0"
+    "validate-npm-package-name": "^4.0.0",
+    "tap": "^15.2.3"
   }
 }
